+++ src/App.js
@@ -1,10 +1,12 @@
 import "./App.css";
 import Navbar from "./components/Navbar";
+import Login from "./Login";
 
 function App() {
   return (
     <div className="App">
-      <Navbar />
+      <Login />
+      {/* <Navbar /> */}
     </div>
   );
 }
