+++ src/components/LoginComponents/Icon.jsx
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export default function Icon({ color, children }) {
+  return <StyledIcon background={color}>{children}</StyledIcon>;
+}
+
+const StyledIcon = styled.div`
+  height: 2.5rem;
+  width: 2.5rem;
+  background: ${(props) => props.background};
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-radius: 4rem;
+  color: white;
+  cursor: pointer;
+  svg {
+    width: 1.5rem;
+    height: 1.5rem;
+  }
+`;
